Cache and reuse handler objects.
This introduces an interesting conundrum.  Now, context handlers need to be careful of which arguments they're processing since they do get reused. That means they could potentially be non-idempotent. We may need to
figure out a way to force the issue, or just hope that no one does anything stupid. I'm not sure which approach is better yet.